Save app configuration and change thumbnail size
Improve handling of configuration. It is now possible to save the app
configuration explicitly (press CTRL+C). The configuration is loaded
automatically at app startup and is used to control a growing number of
app functionalities.

Thumbnail size can now be scaled and this setting can be saved in the
configuration file.