python3/gi: fix main issues
Immagine should now start and be usable with Python 3 and GObject
introspection (gi). Some issues remain. The most annoying are:
(1) images not centered when opened and (2) cannot move the helper
window when in fullscreen mode.

Also the port to Cairo-based drawing is suboptimal to say the least:
Pixbuf-s are still used with conversions via numpy, causing copies and
possibly other issues with binary representation. In particular, there
may be issues with strides being handled correctly. This code should
be refactored, but for now it seems to work.